+++ grid.h
@@ -13,26 +13,28 @@
 #ifndef grid_h
 #define grid_h
 
+#include <stddef.h>
+
 typedef struct grid grid_t;
 
 /**
  * Create a grid of the given max dimensions. Use mem_deref() to destroy.
  */
-grid_t *grid_alloc(int width, int height);
+grid_t *grid_alloc(size_t width, size_t height);
 
 /**
  * Get whatever is at (x, y). Returns NULL if nothing is there.
  */
-void *grid_get(grid_t * grid, int x, int y);
+void *grid_get(grid_t * grid, size_t x, size_t y);
 
 /**
  * Put `userdata` (a pointer to anything) at (x, y) and ref it.
  */
-void grid_put(grid_t * grid, int x, int y, void *obj);
+void grid_put(grid_t * grid, size_t x, size_t y, void *obj);
 
 /**
  * Remove whatever is at (x, y) and deref it.
  */
-void grid_remove(grid_t * grid, int x, int y);
+void grid_remove(grid_t * grid, size_t x, size_t y);
 
 #endif
