+++ mem.h
@@ -10,13 +10,15 @@
 #ifndef mem_h
 #define mem_h
 
+#include <stddef.h>
+
 /**
  * Allocate a chunk of memory with an optional destructor and a refcount.
  *
  * If 'fin' is not NULL it will be called on the final mem_deref. The returned
  * chunk will have a refcount of 1.
  */
-void *mem_alloc(int size, void (*fin) (void *));
+void *mem_alloc(size_t size, void (*fin) (void *));
 
 /**
  * Increment the refcount on a chunk.
@@ -33,12 +35,12 @@ void mem_deref(void *ptr);
 /**
  * Get the number of references on a chunk.
  */
-unsigned int mem_get_refs(void *ptr);
+size_t mem_get_refs(void *ptr);
 
 /**
  * Set a function to call when an error occurs.
  */
-void mem_set_err_handler(void (*handler) (int size));
+void mem_set_err_handler(void (*handler) (size_t size));
 
 /**
  * Return a string describing the last error.
